Updated device list to only show a preview of each device with a power toggle.
Moved device controls into it's own view, which can be accessed by clicking on a device in the device list.
Added color picker back, with only basic functionality of being able to set a color.
Removed the concept of DeviceInfo completely, as the devices endpoint now returns function/variable info, so only the single request is necessary to get the info needed.